requirements.txt
# AI-PM-ToolKit — Root Dependencies
# These are shared across all projects in this repo.
# Each subfolder may have its own requirements.txt with project-specific packages.

# Anthropic Claude API
anthropic>=0.40.0

# Environment variable management
python-dotenv>=1.0.0